Main script
This is the standalone Python script.
To customize it according to one's facility data-management needs a dictionary called DATA is initialized at the beginning of the Python code. Please just modify the dictionary items' values in order to maintain the tool's original funcitonalities. Comments are added in the dictionary for help.